# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Restaurants API",
    "version": "1.0.0",
    "description": "API de restaurantes y comentarios con auth JWT"
  },
  "servers": [
    { "url": "http://localhost:4000", "description": "Local" }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Registro de usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuario creado" },
          "400": { "description": "Error de validación" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login de usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login correcto con token" },
          "401": { "description": "Credenciales incorrectas" }
        }
      }
    },
    "/restaurants": {
      "get": {
        "summary": "Lista todos los restaurantes",
        "responses": {
          "200": { "description": "Array de restaurantes" }
        }
      },
      "post": {
        "summary": "Crea restaurante",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Restaurant"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Restaurante creado" },
          "401": { "description": "No autenticado" }
        }
      }
    },
    "/restaurants/{id}": {
      "get": {
        "summary": "Obtiene un restaurante",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Restaurante" },
          "404": { "description": "No encontrado" }
        }
      },
      "put": {
        "summary": "Edita un restaurante",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Restaurant" } } } },
        "responses": { "200": { "description": "Restaurante actualizado" }, "403": { "description": "No es propietario" }, "404": { "description": "No encontrado" } }
      },
      "delete": {
        "summary": "Elimina un restaurante",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Eliminado" }, "403": { "description": "No es propietario" }, "404": { "description": "No encontrado" } }
      }
    },
    "/restaurants/{id}/comments": {
      "get": {
        "summary": "Lista comentarios de un restaurante",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Lista de comentarios" } }
      },
      "post": {
        "summary": "Crea comentario en un restaurante",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } } },
        "responses": { "201": { "description": "Comentario creado" }, "401": { "description": "No autenticado" } }
      }
    },
    "/comments/{id}": {
      "put": {
        "summary": "Edita comentario",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Comentario editado" }, "403": { "description": "No es propietario" }, "404": { "description": "No encontrado" } }
      },
      "delete": {
        "summary": "Elimina comentario",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Eliminado" }, "403": { "description": "No es propietario" }, "404": { "description": "No encontrado" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Restaurant": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "address": { "type": "string" },
          "description": { "type": "string" },
          "cuisine": { "type": "string" },
          "phone": { "type": "string" },
          "image": { "type": "string" },
          "openingHours": { "type": "string" },
          "lat": { "type": "number" },
          "lng": { "type": "number" }
        },
        "required": ["name", "address", "description"]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "text": { "type": "string" },
          "rating": { "type": "integer", "minimum": 0, "maximum": 5 }
        },
        "required": ["text"]
      }
    }
  }
}